Introduce `go` extension for plugin-gradle
