+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/FormatExtension.java
@@ -957,33 +957,6 @@ public class FormatExtension {
 		return new EclipseWtpConfig(type, version);
 	}
 
-	public class GofmtConfig {
-		GofmtFormatStep stepCfg;
-
-		public GofmtConfig(String version) {
-			stepCfg = (GofmtFormatStep) GofmtFormatStep.withVersion(version).create();
-			addStep(createStep());
-		}
-
-		public GofmtConfig withGoExecutable(String pathToGo) {
-			stepCfg = stepCfg.withGoExecutable(pathToGo);
-			replaceStep(createStep());
-			return this;
-		}
-
-		private FormatterStep createStep() {
-			return stepCfg.create();
-		}
-	}
-
-	public GofmtConfig gofmt() {
-		return new GofmtConfig(GofmtFormatStep.defaultVersion());
-	}
-
-	public GofmtConfig gofmt(String version) {
-		return new GofmtConfig(version);
-	}
-
 	/**
 	 * <pre>
 	 * spotless {
